Refactor code. Add comments to functions
